lib, zebra: fix TE link-params
some of the existing constraints appear to be arbitrary and do not
respect what we see in the RFC. The following changes were made:
* max-bw: represents the actual link bandwidth. MUST be less or equal
  than the bandwidth specified in zebra (after the appropriate unit
  conversion)
* max-rsv-bw: not constrained - specifically stated that can be higher
  than max-bw for over-subscription purposes
* unres-bw: MUST be less or equal than the max-rsv-bw; can be higher
  than the max-bw.

When link-parameters are enabled on an interface, the current code
automatically sets some parameters - namely the metric, max-bw,
max-rsv-bw, and unres-bw for all priorities; I am not sure what was
the rationale behind this. I simplified the code by not setting any
parameters unless they were explicitly set by the user; this avoids
the need to track what was set manually and simplifies the logic that
enforces the constraints above.

Some negative commands were missing, namely all the ones related to
bandwidth: 'no max-bw [BANDWIDTH]', 'no max-rsv-bw [BANDWIDTH]',
'no unres-bw [(1-7) [BANDWIDTH]]'. These were added.

A change of the bandwidth parameter in zebra was not being propagated to
the link parameters unless those were cycled with no enable + enable.
This has been corrected.

As part of this process, we also modify the value of max-bw in case it
was set and higher than the new link bandwidth; in this case we also
fire a log and print a NOTICE in vtysh.

Signed-off-by: Emanuele Di Pascale <emanuele@voltanet.io>